Some small cleanups and permission checks

+++ modules/commands.py
@@ -12,43 +12,50 @@ from utils.logging_utils import LoggingUtils
 class Commands:
     @staticmethod
     async def faculty_roles(ctx: SlashInteraction):
-        options = []
-        with DB.cursor() as cur:
-            cur.execute("SELECT * FROM faculties order by title")
-            for row in cur.fetchall():
-                emoji = None
-                try:
-                    for e in ctx.guild.emojis:
-                        if e.name.lower() == row['icon'].lower():
-                            emoji = e
-                except:
-                    pass
-                options.append(
-                    SelectOption(
-                        row['title'].strip(),
-                        row['role'],
-                        emoji=PartialEmoji(
-                            name=emoji.name if emoji else None,
-                            id=emoji.id if emoji else None
-                        ) if emoji else None
+        author_permission: Permissions = ctx.author.permissions_in(ctx.channel)
+        if author_permission.administrator:
+            await ctx.reply(content='Creating selector', ephemeral=True, delete_after=1)
+            options = []
+            with DB.cursor() as cur:
+                cur.execute("SELECT * FROM faculties order by title")
+                for row in cur.fetchall():
+                    emoji = None
+                    try:
+                        for e in ctx.guild.emojis:
+                            if e.name.lower() == row['icon'].lower():
+                                emoji = e
+                    except:
+                        pass
+                    options.append(
+                        SelectOption(
+                            row['title'].strip(),
+                            row['role'],
+                            emoji=PartialEmoji(
+                                name=emoji.name if emoji else None,
+                                id=emoji.id if emoji else None
+                            ) if emoji else None
+                        )
                     )
-                )
 
-        options = SelectMenu(
-            options=options,
-            custom_id='facultiesSelect',
-            max_values=1,
-            placeholder='Lūdzu izvēlies savu fakultāti'
-        )
-        msg = await ctx.channel.send(
-            content="Izvēlies savu fakultāti (ja esi pārstāvējis vairākas, izvēlies pēdējo vai arī kurai jūties vairāk piederīgs)",
-            components=[options]
-        )
-        with DB.cursor() as cur:
-            cur.execute("INSERT INTO messages (channel, message, type) VALUES (?, ?, ?)", (msg.channel.id, msg.id, 'facultiesSelect'))
+            options = SelectMenu(
+                options=options,
+                custom_id='facultiesSelect',
+                max_values=1,
+                placeholder='Lūdzu izvēlies savu fakultāti'
+            )
+            msg = await ctx.channel.send(
+                content="Izvēlies savu fakultāti (ja esi pārstāvējis vairākas, izvēlies pēdējo vai arī kurai jūties vairāk piederīgs)",
+                components=[options]
+            )
+            with DB.cursor() as cur:
+                cur.execute("INSERT INTO messages (channel, message, type) VALUES (?, ?, ?)", (msg.channel.id, msg.id, 'facultiesSelect'))
+        else:
+            await ctx.reply(content='You need to an administrator to do that', ephemeral=True, delete_after=1)
+
 
     @staticmethod
     async def on_faculty_role(ctx: MessageInteraction, selected: SelectOption):
+        author_permission: Permissions = ctx.author.permissions_in(ctx.channel)
         faculty_roles = []
         user = DB.get_user(ctx.author)
         with DB.cursor() as cur:
@@ -61,7 +68,6 @@ class Commands:
         if len(user_roles) == 1:
             user['faculty_id'] = user_roles[0].id
             user['last_faculty_change'] = datetime.now()-timedelta(20)
-        author_permission: Permissions = ctx.author.permissions_in(ctx.channel)
         if user['faculty_id'] not in [i.id for i in user_roles]:
             if not user['is_faculty_locked']:
                 change_interval = timedelta(days=DB.get_setting('faculty_change_interval_days', 30))
@@ -99,14 +105,20 @@ class Commands:
 
     @staticmethod
     async def faculty_stats(ctx: SlashInteraction):
-        embed = Embed(color=0xeaaa00)
-        embed.title = 'Šobrīdējais fakultāšu pārstāvju skaits serverī'
-        msg = await ctx.channel.send(embed=embed)
+        author_permission: Permissions = ctx.author.permissions_in(ctx.channel)
+        if author_permission.administrator:
+            await ctx.reply(content='Creating stats display', ephemeral=True, delete_after=1)
 
-        with DB.cursor() as cur:
-            cur.execute("INSERT INTO messages (channel, message, type) VALUES (?, ?, ?)", (msg.channel.id, msg.id, 'facultiesStats'))
+            embed = Embed(color=0xeaaa00)
+            embed.title = 'Šobrīdējais fakultāšu pārstāvju skaits serverī'
+            msg = await ctx.channel.send(embed=embed)
 
-        await Commands.update_faculty_stats(ctx)
+            with DB.cursor() as cur:
+                cur.execute("INSERT INTO messages (channel, message, type) VALUES (?, ?, ?)", (msg.channel.id, msg.id, 'facultiesStats'))
+
+            await Commands.update_faculty_stats(ctx)
+        else:
+            await ctx.reply(content='You need to an administrator to do that', ephemeral=True, delete_after=1)
 
     @staticmethod
     async def update_faculty_stats(ctx):
@@ -135,40 +147,53 @@ class Commands:
 
     @staticmethod
     async def send_post(ctx: SlashInteraction, channel: discord.TextChannel, message: str):
-        if isinstance(channel, discord.TextChannel):
-            await channel.send(message.replace('\\n', '\n'))
-            await ctx.send(':thumbsup:', ephemeral=True, delete_after=1)
+        author_permission: Permissions = ctx.author.permissions_in(ctx.channel)
+        if author_permission.administrator:
+            if isinstance(channel, discord.TextChannel):
+                await channel.send(message.replace('\\n', '\n'))
+                await ctx.send(':thumbsup:', ephemeral=True, delete_after=1)
+            else:
+                await ctx.reply('You must provide text channel', ephemeral=True, delete_after=2)
         else:
-            await ctx.reply('You must provide text channel', ephemeral=True, delete_after=2)
+            await ctx.reply(content='You need to an administrator to do that', ephemeral=True, delete_after=1)
 
     @staticmethod
-    async def edit_post(ctx: SlashInteraction, channel: discord.TextChannel, message_id: int, message: str):
-        if isinstance(channel, discord.TextChannel):
-            try:
-                msg = channel.get_partial_message(message_id)
-                if msg:
-                    try:
-                        await msg.edit(content=message.replace('\\n', '\n'))
-                        await ctx.reply(':thumbsup:', ephemeral=True, delete_after=1)
-                    except Exception as e:
-                        print(e)
-                        await ctx.reply('Could not edit message with provided id', ephemeral=True, delete_after=2)
-                else:
+    async def edit_post(ctx: SlashInteraction, channel_id: int, message_id: int, message: str):
+        author_permission: Permissions = ctx.author.permissions_in(ctx.channel)
+        if author_permission.administrator:
+            channel: discord.TextChannel = ctx.guild.get_channel(channel_id)
+            if isinstance(channel, discord.TextChannel):
+                try:
+                    msg = channel.get_partial_message(message_id)
+                    if msg:
+                        try:
+                            await msg.edit(content=message.replace('\\n', '\n'))
+                            await ctx.reply(':thumbsup:', ephemeral=True, delete_after=1)
+                        except Exception as e:
+                            print(e)
+                            await ctx.reply('Could not edit message with provided id', ephemeral=True, delete_after=2)
+                    else:
+                        await ctx.reply('Could not find message with provided id', ephemeral=True, delete_after=2)
+                except Exception as e:
+                    print(e)
                     await ctx.reply('Could not find message with provided id', ephemeral=True, delete_after=2)
-            except Exception as e:
-                print(e)
-                await ctx.reply('Could not find message with provided id', ephemeral=True, delete_after=2)
+            else:
+                await ctx.reply('You must provide text channel', ephemeral=True, delete_after=2)
         else:
-            await ctx.reply('You must provide text channel', ephemeral=True, delete_after=2)
+            await ctx.reply(content='You need to an administrator to do that', ephemeral=True, delete_after=1)
 
     @staticmethod
     async def rules_message(ctx: SlashInteraction):
-        await ctx.reply('Creating rules message:', ephemeral=True, delete_after=2)
-        msg = await ctx.channel.send('**=== NOTEIKUMI ===**')
-        with DB.cursor() as cur:
-            cur.execute("INSERT INTO messages (channel, message, type) VALUES (?, ?, ?)", (msg.channel.id, msg.id, 'rules'))
+        author_permission: Permissions = ctx.author.permissions_in(ctx.channel)
+        if author_permission.administrator:
+            await ctx.reply('Creating rules message:', ephemeral=True, delete_after=2)
+            msg = await ctx.channel.send('**=== NOTEIKUMI ===**')
+            with DB.cursor() as cur:
+                cur.execute("INSERT INTO messages (channel, message, type) VALUES (?, ?, ?)", (msg.channel.id, msg.id, 'rules'))
 
-        await Commands.update_rules_message(ctx)
+            await Commands.update_rules_message(ctx)
+        else:
+            await ctx.reply(content='You need to an administrator to do that', ephemeral=True, delete_after=1)
 
     @staticmethod
     async def update_rules_message(ctx: SlashInteraction):
@@ -192,32 +217,44 @@ class Commands:
 
     @staticmethod
     async def rules_add(ctx: SlashInteraction, msg: str):
-        await ctx.reply('Updating rules:', ephemeral=True, delete_after=2)
-        with DB.cursor() as cur:
-            cur.execute("SELECT max(id) as mid FROM rules")
-            mid = cur.fetchone()['mid']
-            if not mid:
-                mid = 0
-            cur.execute("INSERT INTO rules VALUES "
-                        "(?, ?)", (mid+1, msg,))
+        author_permission: Permissions = ctx.author.permissions_in(ctx.channel)
+        if author_permission.administrator:
+            await ctx.reply('Updating rules:', ephemeral=True, delete_after=2)
+            with DB.cursor() as cur:
+                cur.execute("SELECT max(id) as mid FROM rules")
+                mid = cur.fetchone()['mid']
+                if not mid:
+                    mid = 0
+                cur.execute("INSERT INTO rules VALUES "
+                            "(?, ?)", (mid+1, msg,))
 
-        await Commands.update_rules_message(ctx)
+            await Commands.update_rules_message(ctx)
+        else:
+            await ctx.reply(content='You need to an administrator to do that', ephemeral=True, delete_after=1)
 
     @staticmethod
     async def rules_remove(ctx: SlashInteraction, rid: int):
-        await ctx.reply('Updating rules:', ephemeral=True, delete_after=2)
-        with DB.cursor() as cur:
-            cur.execute('DELETE FROM rules WHERE id=?', (rid,))
-            cur.execute('UPDATE rules SET id = id-1 WHERE id > ?', (rid,))
+        author_permission: Permissions = ctx.author.permissions_in(ctx.channel)
+        if author_permission.administrator:
+            await ctx.reply('Updating rules:', ephemeral=True, delete_after=2)
+            with DB.cursor() as cur:
+                cur.execute('DELETE FROM rules WHERE id=?', (rid,))
+                cur.execute('UPDATE rules SET id = id-1 WHERE id > ?', (rid,))
 
-        await Commands.update_rules_message(ctx)
+            await Commands.update_rules_message(ctx)
+        else:
+            await ctx.reply(content='You need to an administrator to do that', ephemeral=True, delete_after=1)
 
     @staticmethod
     async def rules_reorder(ctx: SlashInteraction, rid: int, new_id: int):
-        await ctx.reply('Updating rules:', ephemeral=True, delete_after=2)
-        with DB.cursor() as cur:
+        author_permission: Permissions = ctx.author.permissions_in(ctx.channel)
+        if author_permission.administrator:
+            await ctx.reply('Updating rules:', ephemeral=True, delete_after=2)
+            with DB.cursor() as cur:
 
-            cur.execute('UPDATE rules SET id = id+1 WHERE id >= ?', (new_id,))
-            cur.execute('UPDATE rules SET id = ? WHERE id = ?', (new_id, rid + 1))
+                cur.execute('UPDATE rules SET id = id+1 WHERE id >= ?', (new_id,))
+                cur.execute('UPDATE rules SET id = ? WHERE id = ?', (new_id, rid + 1))
 
-        await Commands.update_rules_message(ctx)
+            await Commands.update_rules_message(ctx)
+        else:
+            await ctx.reply(content='You need to an administrator to do that', ephemeral=True, delete_after=1)
